Extract reusable filter elements to a mixin
Statistic-based filtering will reuse it every time.
